+++ errors/XlsxDomErrorsImpl.ts
@@ -1,4 +1,4 @@
-﻿/** Common Excel file XML DOM error checking/throwing functions
+﻿/** Common XLSX XML file DOM error checking/throwing functions
  * @author TeamworkGuy2
  * @since 2016-5-27
  */
@@ -7,11 +7,11 @@ module XlsxDomErrorsImpl {
 
 
     export function missingNode(nodeName: string) {
-        return new Error("Error reading Excel template, missing required '" + nodeName + "' node");
+        return new Error("Error reading XLSX template, missing required '" + nodeName + "' node");
     }
 
 
-    export function expectNode(node: Element, expectedNodeName: string, parentNodeName: string, idx?: number, size?: number) {
+    export function expectNode(node: { tagName: string }, expectedNodeName: string, parentNodeName: string, idx?: number, size?: number) {
         if (node.tagName !== expectedNodeName) {
             throw unexpectedNode(node.tagName, expectedNodeName, parentNodeName, idx, size);
         }
@@ -19,7 +19,7 @@ module XlsxDomErrorsImpl {
 
 
     export function unexpectedNode(badNodeName: string, expectedNodeName?: string, parentNodeName?: string, idx?: number, size?: number) {
-        return new Error("Error reading Excel template, unexpected '" + badNodeName + "' node" +
+        return new Error("Error reading XLSX template, unexpected '" + badNodeName + "' node" +
             (expectedNodeName ? ", expected only '" + expectedNodeName + "' nodes" : "") +
             (parentNodeName ? ", of parent node '" + parentNodeName + "'" : "") +
             (idx || size ? (idx ? ", index=" + idx : "") + (size ? ", size=" + size : "") : "")