+++ files/XlsxFileType.ts
@@ -29,6 +29,12 @@ class XlsxFileTypeImpl implements OpenXmlIo.XlsxFileType {
         this.pathTemplateToken = pathTemplateToken;
     }
 
+
+    public static getXmlFilePath(sheetNum: number, info: OpenXmlIo.XlsxFileType) {
+        // TODO the path template token may not be a sheet number, could be a resource identifier (i.e. an image or item prop number)
+        return (info.pathIsTemplate ? info.xlsxFilePath.split(info.pathTemplateToken).join(<string><any>sheetNum) : info.xlsxFilePath);
+    }
+
 }
 
 export = XlsxFileTypeImpl;